testproject/requirements.txt
Django==1.3.1
amqplib==1.0.2
anyjson==0.3.1
celery==2.4.4
django-celery==2.4.2
django-picklefield==0.1.9
importlib==1.0.2
kombu==1.5.1
ordereddict==1.1
python-dateutil==1.5
wsgiref==0.1.2
